refactor(web): formatRelativeTime 순수 함수 추출 + 테스트 5건 (#140)
DashboardPage 인라인 함수를 lib/timeFormat.ts 로 추출.
now 파라미터 주입으로 테스트 결정성 확보.

Refs: #139
